requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
cloudpickle==2.1.0
coreapi==2.3.3
coreschema==0.0.4
dask==2022.8.1
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
flake8==5.0.4
fsspec==2022.7.1
greenlet==1.1.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
locket==1.0.0
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.23.2
packaging==21.3
pandas==1.4.3
partd==1.3.0
progress==1.6
progressbar==2.5
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
SQLAlchemy==1.4.40
sqlparse==0.4.2
toolz==0.12.0
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.12
